Add station list sorting (#37)
* Work on stations sorting/order
* Simpler station regex search:
Don't use other language fields for search (Swedish, English). They are not displayed, or even returned from HTTP API, so it results could be confusing.
Don't use match count for sorting. Also confusing if you want to sort the results by some other fields as well. If the regex   matches, it's in the results. Previously if the pattern was matched in multiple times for some station, it had a higher preference in results.
* Remove empty station search terms
* Send all station search terms
* Reset station list page to 0 when sorting changes:
Otherwise the results wouldn't make sense currently. Other option would be to somehow get to current page on new order but much more complex.